requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-weaviate>=0.0.3
langgraph>=0.1.0
weaviate-client>=4.5.0
openai>=1.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
ragas>=0.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pandas>=2.0.0
spacy>=3.7.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn>=0.29.0
